26-FE: Move router constants into separate file
