+++ Okane.Client/src/features/auth/useAuthStore.ts
@@ -3,12 +3,14 @@ import { computed, ref } from 'vue'
 import { defineStore } from 'pinia'
 
 // Internal
+import { ROUTE_NAME } from '@/shared/services/router/router.constants'
+
 import type { AuthenticateResponse } from '@/features/auth/auth.types'
 import type { User } from '@/features/users/user.types'
 import type { Timeout } from '@/shared/types/shared.type'
 
 import { apiClient } from '@/shared/services/apiClient/apiClient.service'
-import { ROUTE_NAME, router } from '@/shared/services/router/router.service'
+import { router } from '@/shared/services/router/router.service'
 
 import { getJWTTokenPayload } from '@/features/auth/auth.utils'
