+++ Okane.Client/src/shared/components/NavBar.vue
@@ -3,7 +3,7 @@
 import { RouterLink } from 'vue-router'
 
 // Internal
-import { ROUTE_NAME } from '@/shared/services/router/router.service'
+import { ROUTE_NAME } from '@/shared/services/router/router.constants'
 
 import { useAuthStore } from '@/features/auth/useAuthStore'
