+++ Okane.Client/src/shared/pages/LoginPage.vue
@@ -7,10 +7,11 @@ import AuthForm from '@/features/auth/AuthForm.vue'
 import Heading from '@/shared/components/Heading.vue'
 import PageLayout from '@/shared/layouts/PageLayout.vue'
 
+import { ROUTE_NAME } from '@/shared/services/router/router.constants'
+
 import type { AuthFormState } from '@/features/auth/auth.types'
 
 import { useAuthStore } from '@/features/auth/useAuthStore'
-import { ROUTE_NAME } from '@/shared/services/router/router.service'
 
 const authStore = useAuthStore()
 const router = useRouter()
