+++ Okane.Client/src/shared/services/router/router.utils.ts
@@ -2,7 +2,8 @@
 import type { RouteLocationNamedRaw } from 'vue-router'
 
 // Internal
-import type { ROUTE_NAME } from '@/shared/services/router/router.service'
+import { ROUTE_NAME } from '@/shared/services/router/router.constants'
+
 import { router } from '@/shared/services/router/router.service'
 
 export function getURLByRouteName(routeName: ROUTE_NAME, options?: RouteLocationNamedRaw) {
